// apps/hanwha-ocean/src/pages/crane-detail/model/types.ts
export interface CmmsTrolleyUnit {
  runFault: RunFaultStatus;
  joystickStep: string;
  speedRef: number;
  actSpeed: number;
  current: number;
  driveFault: string;
  position: number;
  // 상태 비트
  mechStopRelay: OnOff;
  driveFaultBit: OnOff;
  driveMainCb: OnOff;
  driveFanCb: OnOff;
  fieldCb: OnOff;
  driveMainMc: OnOff;
  driveFanMc: OnOff;
  driveFieldMc: OnOff;
  driveMainFuseBlown: OnOff;
  driveFieldFuseBlown: OnOff;
  driveOutputFuseBlown: OnOff;
  motor12FieldMonitor: OnOff;
  motor34FieldMonitor: OnOff;
  motor1Overtemp: OnOff;
  motor2Overtemp: OnOff;
  motor3Overtemp: OnOff;
  motor4Overtemp: OnOff;
  eStopPb: OkNg;
}